Give every API model its own file
