+++ client/app/core/table-name/table-name.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnChanges, SimpleChanges } from '@angular/core';
-import { TableName } from '../../common/table-name.class';
+import { TableName } from '../../common/table-name';
 
 @Component({
     selector: 'table-name',
