+++ client/app/tables/api-data-source/api-data-source.ts
@@ -7,8 +7,8 @@ import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { Observable } from 'rxjs/Rx';
 import { Subscription } from 'rxjs/Subscription';
 import { Filter, SqlRow, TableHeader, TableMeta } from '../../common/api';
+import { PaginatedResponse } from '../../common/api';
 import { DATE_FORMAT, DATETIME_FORMAT } from '../../common/constants';
-import { PaginatedResponse } from '../../common/responses';
 import { ContentRequest } from '../../core/table/content-request';
 import { TableService } from '../../core/table/table.service';
 import { FilterManagerComponent } from '../filter-manager/filter-manager.component';
