+++ client/app/tables/table-host/table-host.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params } from '@angular/router';
-import { TableName } from '../../common/table-name.class';
+import { TableName } from '../../common/table-name';
 
 @Component({
     templateUrl: 'table-host.component.html',
