+++ common/api/constraint-type.ts
@@ -1,0 +1,5 @@
+
+/**
+ * The types of constraints currently supported by Helium
+ */
+export type ConstraintType = 'primary' | 'foreign' | 'unique';
