+++ common/api/default-value.ts
@@ -1,0 +1,3 @@
+import { SpecialDefaultValue } from '.';
+
+export type DefaultValue = string | number | boolean | null | Date | SpecialDefaultValue;
