+++ common/api/error-response.ts
@@ -1,11 +1,6 @@
+
 export interface ErrorResponse {
     message: string;
     error?: any;
     relevantInput?: any;
-}
-
-export interface PaginatedResponse<T> {
-    size: number;
-    data: T;
-    totalRows: number;
 }
