+++ common/api/master-table-name.ts
@@ -1,0 +1,9 @@
+import { BaseTableName } from '.';
+import { TableName } from '../table-name';
+
+export interface MasterTableName extends BaseTableName {
+    /**
+     * Any part tables that belong to this master table.
+     */
+    parts: TableName[];
+}
