+++ common/api/special-default-value.ts
@@ -1,0 +1,2 @@
+
+export interface SpecialDefaultValue { constantName: string; }
