+++ common/api/sql-row.ts
@@ -1,0 +1,4 @@
+
+export interface SqlRow {
+    [columnName: string]: any;
+}
