+++ common/api/table-data-type.ts
@@ -1,0 +1,2 @@
+
+export type TableDataType = 'string' | 'integer' | 'float' | 'date' | 'datetime' | 'boolean' | 'enum' | 'blob';
