+++ common/api/table-insert.ts
@@ -1,5 +1,4 @@
-
-import { SqlRow } from './api';
+import { SqlRow } from '.';
 
 export interface TableInsert {
     [tableName: string]: SqlRow[];
