+++ common/api/table-tier.ts
@@ -1,0 +1,6 @@
+
+/**
+ * DataJoint table tier. See http://docs.datajoint.io/data-definition/Data-tiers.html
+ * for more.
+ */
+export type TableTier = 'lookup' | 'manual' | 'imported' | 'computed' | 'hidden';
