+++ server/src/db/schema-input.validator.ts
@@ -4,9 +4,9 @@ import * as JoiDateExtensions from 'joi-date-extensions';
 import * as _ from 'lodash';
 import * as moment from 'moment';
 import { TableHeader } from '../common/api';
+import { TableInsert } from '../common/api/table-insert';
 import { DATE_FORMAT, DATETIME_FORMAT } from '../common/constants';
-import { TableInsert } from '../common/table-insert.interface';
-import { TableName } from '../common/table-name.class';
+import { TableName } from '../common/table-name';
 import { ErrorCode } from '../routes/api/error-code';
 import { ValidationError } from '../routes/api/validation-error';
 import { SchemaDao } from './schema.dao';
